Omstrukturer pakker og klasser, og begynner å legge til tester
