+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/Tidspunkt.kt
@@ -260,58 +260,11 @@ data class MånedTidspunkt(
     override fun erDag() = false
 }
 
-data class Tidsrom(
-    override val start: Tidspunkt,
-    override val endInclusive: Tidspunkt
-) : Iterable<Tidspunkt>,
-    ClosedRange<Tidspunkt> {
-
-    override fun iterator(): Iterator<Tidspunkt> =
-        if (start.erDag() && !endInclusive.erDag())
-            TidspunktIterator(start, endInclusive.tilSisteDagIMåneden())
-        else if (!start.erDag() && endInclusive.erDag())
-            TidspunktIterator(start.tilFørsteDagIMåneden(), endInclusive)
-        else
-            TidspunktIterator(start, endInclusive)
-
-    override fun toString(): String =
-        "$start - $endInclusive"
-
-    companion object {
-        private class TidspunktIterator(
-            val startTidspunkt: Tidspunkt,
-            val tilOgMedTidspunkt: Tidspunkt
-        ) : Iterator<Tidspunkt> {
-
-            private var gjeldendeTidspunkt = startTidspunkt.somEndelig()
-
-            override fun hasNext() =
-                gjeldendeTidspunkt.neste() <= tilOgMedTidspunkt.neste().somEndelig()
-
-            override fun next(): Tidspunkt {
-                val next = gjeldendeTidspunkt
-                gjeldendeTidspunkt = gjeldendeTidspunkt.neste()
-
-                return if (next == tilOgMedTidspunkt.somEndelig())
-                    tilOgMedTidspunkt
-                else if (next == startTidspunkt.somEndelig())
-                    startTidspunkt
-                else
-                    next
-            }
-        }
-
-        val NULL = Tidspunkt.med(LocalDate.now())..Tidspunkt.med(LocalDate.now().minusDays(2))
-    }
-}
-
-operator fun Tidspunkt.rangeTo(tilOgMed: Tidspunkt) =
-    Tidsrom(this, tilOgMed)
-
 fun Iterable<Tidspunkt>.minsteEllerUendelig() = this.reduce { acc, neste ->
     if (acc.erUendeligLengeSiden() && neste.erUendeligLengeSiden())
         minOf(acc.somEndelig(), neste.somEndelig()).somUendeligLengeSiden()
     else if (acc.erUendeligLengeSiden() || neste.erUendeligLengeSiden())
+    // Må gå til tidspunktet før den endelig perioden for å fange opp at den ikke er uendelig
         minOf(acc.somEndelig(), neste.somEndelig()).forrige().somUendeligLengeSiden()
     else
         minOf(acc, neste).somEndelig()
@@ -321,6 +274,7 @@ fun Iterable<Tidspunkt>.størsteEllerUendelig() = this.reduce { acc, neste ->
     if (acc.erUendeligLengeTil() && neste.erUendeligLengeTil())
         maxOf(acc.somEndelig(), neste.somEndelig()).somUendeligLengeTil()
     else if (acc.erUendeligLengeTil() || neste.erUendeligLengeTil())
+    // Må gå til tidspunktet etter den endelig perioden for å fange opp at den ikke er uendelig
         maxOf(acc.somEndelig(), neste.somEndelig()).neste().somUendeligLengeTil()
     else maxOf(acc, neste).somEndelig()
 }
