+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/util/CharTidslinjeTest.kt
@@ -1,0 +1,55 @@
+package no.nav.familie.ba.sak.kjerne.tidslinje.util
+
+import no.nav.familie.ba.sak.kjerne.tidslinje.komprimer
+import no.nav.familie.ba.sak.kjerne.tidslinje.rangeTo
+import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.Test
+
+class CharTidslinjeTest {
+    @Test
+    fun testEnkelCharTidsline() {
+        val tegn = "---------------"
+        val charTidslinje = CharTidslinje(tegn, jan(2020))
+        Assertions.assertEquals(tegn.length, charTidslinje.perioder().size)
+
+        val perioder = charTidslinje.komprimer().perioder()
+        Assertions.assertEquals(1, perioder.size)
+
+        val periode = perioder.first()
+        Assertions.assertEquals(jan(2020), periode.fom)
+        Assertions.assertEquals(mar(2021), periode.tom)
+        Assertions.assertEquals('-', periode.innhold)
+    }
+
+    @Test
+    fun testUendeligCharTidslinje() {
+        val tegn = "<--->"
+        val charTidslinje = CharTidslinje(tegn, jan(2020))
+
+        Assertions.assertEquals(tegn.length, charTidslinje.perioder().size)
+
+        val perioder = charTidslinje.komprimer().perioder()
+
+        Assertions.assertEquals(1, perioder.size)
+        val periode = perioder.first()
+        Assertions.assertTrue(periode.fom.erUendeligLengeSiden())
+        Assertions.assertTrue(periode.tom.erUendeligLengeTil())
+        Assertions.assertEquals(jan(2020), periode.fom.somEndelig())
+        Assertions.assertEquals(mai(2020), periode.tom.somEndelig())
+        Assertions.assertEquals('-', periode.innhold)
+    }
+
+    @Test
+    fun testSammensattTidsline() {
+        val tegn = "aabbbbcdddddda"
+        val charTidslinje = CharTidslinje(tegn, jan(2020))
+        Assertions.assertEquals(tegn.length, charTidslinje.perioder().size)
+        val perioder = charTidslinje.komprimer().perioder().toList()
+        Assertions.assertEquals(5, perioder.size)
+        Assertions.assertEquals((jan(2020)..feb(2020)).med('a'), perioder[0])
+        Assertions.assertEquals((mar(2020)..jun(2020)).med('b'), perioder[1])
+        Assertions.assertEquals((jul(2020)..jul(2020)).med('c'), perioder[2])
+        Assertions.assertEquals((aug(2020)..jan(2021)).med('d'), perioder[3])
+        Assertions.assertEquals((feb(2021)..feb(2021)).med('a'), perioder[4])
+    }
+}
