+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/util/TidslinjePrint.kt
@@ -1,0 +1,9 @@
+package no.nav.familie.ba.sak.kjerne.tidslinje.util
+
+import no.nav.familie.ba.sak.kjerne.tidslinje.Tidslinje
+
+fun Iterable<Tidslinje<*>>.print() = this.forEach { it.print() }
+fun Tidslinje<*>.print() {
+    println("${this.tidsrom()} ${this.javaClass.name}")
+    this.perioder().forEach { println(it) }
+}
